requirements.txt
streamlit
pandas
numpy
googlemaps
matplotlib
folium
streamlit-folium
python-dotenv